WIP
the one edge per polyline is now broken since the addition of intersection point on polyline
that are of degree at least 4 make them final points. Maybe adding some checks on the polyline
to detect earlier the intersection would solve it ? At least this would allow to detect intersection
of more that 2 triangles along an edge.
Also with the current approach an edge that was selected for an edge per polyline might no longer
exist (split) and we should update that.